CarIssueController

+++ app/Http/Controllers/api/BranchController.php
@@ -9,17 +9,20 @@ use App\Http\Resources\BranchResource;
 use App\Models\Branch;
 use App\Services\BranchService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class BranchController extends Controller
 {
 
-    public function __construct(protected BranchService $branchService){}
+    public function __construct(protected BranchService $branchService)
+    {
+    }
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        $branches = $this->branchService->all(request()->all(),[]);
+        $branches = $this->branchService->all(request()->all(), []);
 
         return $this->res(BranchResource::collection($branches));
 
@@ -30,6 +33,8 @@ class BranchController extends Controller
      */
     public function store(StoreBranchRequest $request)
     {
+        Gate::authorize('create', Branch::class);
+
         $data = $request->validated();
 
         $branch = $this->branchService->createBranch($data);
@@ -42,6 +47,8 @@ class BranchController extends Controller
      */
     public function show(Branch $branch)
     {
+        Gate::authorize('view', $branch);
+
         return $this->res(BranchResource::make($branch));
     }
 
@@ -50,9 +57,11 @@ class BranchController extends Controller
      */
     public function update(UpdateBranchRequest $request, Branch $branch)
     {
-         $data = $request->validated();
+        Gate::authorize('update', $branch);
 
-         $this->branchService->updateBranch($branch, $data);
+        $data = $request->validated();
+
+        $this->branchService->updateBranch($branch, $data);
 
         return $this->res(BranchResource::make($branch));
     }
@@ -62,6 +71,7 @@ class BranchController extends Controller
      */
     public function destroy(Branch $branch)
     {
+        Gate::authorize('delete', $branch);
         $this->branchService->deleteBranch($branch);
         return $this->res();
     }
